Fixing PDF creation
Fixing this was easier than I though. Once I figured out that it was doubling up the path, was easily able to fix it. Could refactor all uses of card_set to more accurately deal with subfolders but... the PDF creation was the only spot where it matters, so I just kludged in an os.path.basename() and was done with it.